Moved web configuration to WebConfig from ApplicationConfig
